Validate path in `HttpArtifactRepository.list_artifacts` (#10585)
