+++ mlflow/server/handlers.py
@@ -102,11 +102,10 @@ from mlflow.tracking._tracking_service.registry import TrackingStoreRegistry
 from mlflow.tracking.registry import UnsupportedModelRegistryStoreURIException
 from mlflow.utils.file_utils import local_file_uri_to_path
 from mlflow.utils.mime_type_utils import _guess_mime_type
-from mlflow.utils.os import is_windows
 from mlflow.utils.promptlab_utils import _create_promptlab_run_impl
 from mlflow.utils.proto_json_utils import message_to_json, parse_dict
 from mlflow.utils.string_utils import is_string_type
-from mlflow.utils.uri import is_file_uri, is_local_uri
+from mlflow.utils.uri import is_file_uri, is_local_uri, validate_path_is_safe
 from mlflow.utils.validation import _validate_batch_log_api_req
 
 _logger = logging.getLogger(__name__)
@@ -539,30 +538,6 @@ def _disable_if_artifacts_only(func):
         return func(*args, **kwargs)
 
     return wrapper
-
-
-_OS_ALT_SEPS = [sep for sep in [os.sep, os.path.altsep] if sep is not None and sep != "/"]
-
-
-def validate_path_is_safe(path):
-    """
-    Validates that the specified path is safe to join with a trusted prefix. This is a security
-    measure to prevent path traversal attacks.
-    A valid path should:
-        not contain separators other than '/'
-        not contain .. to navigate to parent dir in path
-        not be an absolute path
-    """
-    if is_file_uri(path):
-        path = local_file_uri_to_path(path)
-    if (
-        any((s in path) for s in _OS_ALT_SEPS)
-        or ".." in path.split("/")
-        or pathlib.PureWindowsPath(path).is_absolute()
-        or pathlib.PurePosixPath(path).is_absolute()
-        or (is_windows() and len(path) >= 2 and path[1] == ":")
-    ):
-        raise MlflowException(f"Invalid path: {path}", error_code=INVALID_PARAMETER_VALUE)
 
 
 @catch_mlflow_exception
